+++ system/src/documents/ActorSD.mjs
@@ -129,6 +129,12 @@ export default class ActorSD extends foundry.documents.Actor {
 			prototypeToken.actorLink = true;
 		}
 
+		if (data.type === "Vehicle") {
+			prototypeToken.actorLink = true;
+			prototypeToken.bar1 = { attribute: "attributes.hp" };
+			prototypeToken.disposition = 0;
+		}
+
 		const update = {prototypeToken};
 
 		if (!data.img) {
@@ -376,6 +382,10 @@ export default class ActorSD extends foundry.documents.Actor {
 
 
 	async canUseMagicItems() {
+		if (this.type === "Vehicle") {
+			return false;
+		}
+
 		const characterClass = await this.getClass();
 
 		const spellcastingClass =
@@ -626,6 +636,10 @@ export default class ActorSD extends foundry.documents.Actor {
 
 
 	getCalculatedAbilities() {
+		if (this.type === "Vehicle") {
+			return {};
+		}
+
 		const abilities = {};
 
 		for (const ability of CONFIG.SHADOWDARK.ABILITY_KEYS) {
@@ -807,6 +821,10 @@ export default class ActorSD extends foundry.documents.Actor {
 
 
 	async isSpellCaster() {
+		if (this.type === "Vehicle") {
+			return false;
+		}
+
 		const characterClass = await this.getClass();
 
 		const spellcastingClass =
@@ -1256,6 +1274,26 @@ export default class ActorSD extends foundry.documents.Actor {
 			content,
 			rollMode: CONST.DICE_ROLL_MODES.PUBLIC,
 		});
+	}
+
+
+	async vehiclePropertyItems() {
+		const propertyItems = [];
+		for (const uuid of this.system.properties ?? []) {
+			const item = await fromUuid(uuid);
+			if (item) propertyItems.push(item);
+		}
+		return propertyItems;
+	}
+
+
+	async hasVehicleProperty(property) {
+		property = property.slugify();
+		const propertyItems = await this.vehiclePropertyItems();
+		const propertyItem = propertyItems.find(
+			p => p.name.slugify() === property
+		);
+		return propertyItem ? true : false;
 	}
 
 }