+++ system/src/models/_module.mjs
@@ -1,6 +1,7 @@
 
 export {default as PlayerSD} from "./PlayerSD.mjs";
 export {default as NpcSD} from "./NpcSD.mjs";
+export {default as VehicleSD} from "./VehicleSD.mjs";
 
 export {default as AncestrySD} from "./items/AncestrySD.mjs";
 export {default as ArmorSD} from "./items/ArmorSD.mjs";