+++ system/src/sheets/_module.mjs
@@ -3,3 +3,4 @@ export {default as ItemSheetSD} from "./ItemSheetSD.mjs";
 export {default as NpcSheetSD} from "./NpcSheetSD.mjs";
 export {default as PlayerSheetSD} from "./PlayerSheetSD.mjs";
 export {default as LightSheetSD} from "./LightSheetSD.mjs";
+export {default as VehicleSheetSD} from "./VehicleSheetSD.mjs";